pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "awesome-babushka-commons"
version = "0.1.1"
description = "Common utilities and database connections for awesome-babushka projects"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "certifi==2025.7.14",
    "charset-normalizer==3.4.2",
    "dnspython==2.7.0",
    "fastapi==0.116.1",
    "idna==3.10",
    "pydantic[email]==2.11.7",
    "pydantic-core==2.33.2",
    "pymongo==4.13.2",
    "python-jose==3.3.0",
    "requests==2.32.4",
    "sniffio==1.3.1",
    "starlette==0.47.2",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "urllib3==2.5.0",
]

[project.urls]
"Homepage" = "https://github.com/himansu9805/awesome-babushka-commons"
"Bug Tracker" = "https://github.com/himansu9805/awesome-babushka-commons/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 79
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 79
